+++ backend/.env
@@ -1,4 +1,4 @@
-PORT = 3000
+PORT = 5000
 
 # postgres settings
 #  add your own configuration
